fix: Prevent tabbing onto off screen markers
Hide the marker if it's not visible on the current map display (this will hide the marker from viewers, keyboard users and screen readers). Hiding the marker from keyboard users is important so that they cannot tab onto these markers since tabbing onto an off screen marker causes the mapDiv to scroll automatically which shifts the controls of the map and sometimes breaks the map.