fix: Enable scrolling ability in info window
If gesture handling is not set to cooperative in the map creation options, then scrolling in the info window would just zoom the map instead of enabling scrolling in the info window. This fix reverses that and allows scrolling in the info window with the cost that extra steps are needed to scroll the map (CTRL + scroll).