fix: Show hidden marker if it's li elem is clicked
Initially, if the user clicks on a list item in the side bar and the corresponding marker was hidden from view, then the info window would open there correctly, but the marker would remain hidden. This fix ensures that the marker comes back to view since the info window opening automatically ensures the marker is in view.